requirements.txt
psycopg2-binary~=2.9.9
requests~=2.31.0
interactions.py~=5.11.0
discord.py~=2.32
asyncio~=3.8.1